fix(quant): widen calibrate_and_quantize_fp8_async n_elements + guard int overflow
Audit F-A11 (low). The FP8 migrate API took `int n_elements` and callers passed
static_cast<int>(size_t), silently truncating. No current tensor is close
(largest ~778M elems << 2^31), but a >2.1B-element tensor would corrupt with a
wrapped grid + indices. Widen the parameter to int64_t (callers pass the full
size_t) and guard the boundary loudly inside: if n_elements > INT_MAX, log and
skip rather than truncate. The reduction/quantize kernels keep int indexing —
they only ever run for guarded, in-range sizes now.

Validated: build green, FP8 KV oracle tests pass.

+++ src/exec/pre_dequant_phase3_nvfp4_decode.cu
@@ -637,7 +637,7 @@ void QuantPipeline::nvfp4_decode_free_fp16_and_migrate_fp8_(size_t& remaining_bu
                 fp8_offset += e.n_elems;
 
                 calibrate_and_quantize_fp8_async(e.fp16_tensor.data, fp8_buf,
-                                                 static_cast<int>(e.n_elems), d_block_maxes, max_grid,
+                                                 static_cast<int64_t>(e.n_elems), d_block_maxes, max_grid,
                                                  d_absmax, d_scales_all + i, stream);
 
                 Tensor fp8_t(fp8_buf, QType::FP8_E4M3, e.fp16_tensor.ndim, e.fp16_tensor.shape, true);
